deploy + test auction params

+++ contracts/DecentPoems.sol
@@ -225,7 +225,7 @@ contract DecentPoems is ERC721Royalty, Ownable, VRFConsumerBaseV2 {
     function setAuctionParams(
         uint256 auctionDuration,
         uint256 auctionStartPrice,
-        uint32 auctionEndPrice
+        uint256 auctionEndPrice
     ) public onlyOwner {
         _auctionDuration = auctionDuration;
         _auctionStartPrice = auctionStartPrice;
